Update Day 8 - Dictionaries and Maps.py
In the question, it's given that an unknown number of names will be given to query our phone book. The code that was written here will take only N number of inputs for the query which will show an error when the number of queries exceeds N, EOF error.
The code that I have provided will take input9 until the user stops, which is what the question demands.